refactor supervisor classifiers
