--- RSimageprocess/supervise_classifier.cpp
@@ -1,8 +1,0 @@
-//
-//  classifier.cpp
-//  RSimageprocess
-//
-//  Created by ChanningTong on 12/16/24.
-//
-
-#include "supervise_classifier.hpp"
