--- RSimageprocess/unsupervise_classifier.cpp
@@ -1,8 +1,0 @@
-//
-//  unsupervise_classifier.cpp
-//  RSimageprocess
-//
-//  Created by ChanningTong on 12/16/24.
-//
-
-#include "unsupervise_classifier.hpp"
