--- RSimageprocess/unsupervise_classifier.hpp
@@ -1,16 +1,0 @@
-//
-//  unsupervise_classifier.hpp
-//  RSimageprocess
-//
-//  Created by ChanningTong on 12/16/24.
-//
-
-#ifndef unsupervise_classifier_hpp
-#define unsupervise_classifier_hpp
-
-enum class UnsuperviseClassifierType{
-    isodata,
-    kmean
-};
-
-#endif /* unsupervise_classifier_hpp */
